Move dummy mbus messages to shared header and use in MergeThrottler as well
Explicitly override `mbus::Message::getApproxSize()` to return 0 (instead
of the default 1) to avoid mis-counting size between requests and responses.
This has not mattered in practice since we haven't set any size-based
limits (only message count limits), but we should ensure symmetry anyway.
